Change grid columns based on container width
This leverages the new [@container](https://developer.mozilla.org/en-US/docs/Web/CSS/@container) query, which is present in the latest versions of all major browsers, and allows us to dynamically set the column count based on of the width of the parent container.

This means that if a repeater is used within another block that constricts the width (say, a two column layout), this will shrink the number of columns. The target is to give each column a minimum of roughly 400 pixels.

My understanding is that browsers that do not yet support this query will simply ignore it, which will fall back to using the viewport media queries.